+++ PortalLib/src/main/java/ua/dp/stud/StudPortalLib/service/CourseService.java
@@ -23,11 +23,11 @@ public interface CourseService {
 
     List<Course> getCoursesByKindAndType(String kindOfCourse, String coursesType);
 
-    KindOfCourse getKindOfCourseById(Integer id);
+    KindOfCourse getKindOfCourseById(Long id);
 
     void addKindOfCourse(KindOfCourse kindOfCourse);
 
-    void deleteKindOfCourse(Integer id);
+    void deleteKindOfCourse(Long id);
 
     void updateKindOfCourse(KindOfCourse kindOfCourse);
